# openapi.yaml
openapi: 3.0.3
info:
  title: SmartGreenAI - Radish Crop API
  version: 1.3.0
  description: |
    Contrato REST entre la Raspberry Pi (gateway), el backend en la nube y el
    dashboard web. Cubre SYS-01 y define lo que implementan CONF-04 (registro),
    MON-03 (cliente de transmision en la Pi) y MON-04 (ingesta y almacenamiento).

    **v1.1.0 (Sprint 2):** agrega sesiones de usuario (AUTH-01), control de
    acceso por rol (AUTH-02), registro de actuadores (CONF-05) y ultimas
    lecturas (DASH-01).

    **v1.2.0 (Sprint 2):** ajuste de MON-04 para el gateway de la Pi (MON-03).
    `ts` debe traer zona horaria y se guarda en UTC con milisegundos; se
    rechazan horas imposibles (reloj sin sincronizar o del futuro) y la
    validacion es lectura por lectura: una lectura mala ya no tumba el lote.

    **v1.3.0 (Sprint 2):** soporte para los widgets del dashboard (DASH-01).
    Agrega `GET /greenhouses` (viveros que el usuario puede consultar), el campo
    `area` en las ultimas lecturas y el header opcional `X-SG-Background` para
    que el refresco automatico del dashboard no renueve la sesion.

    ## Convenciones que no se negocian

    1. **`null` = sensor caido. `0` = lectura real.** Cualquier campo de medicion
       puede venir en `null`. El backend NUNCA debe convertir `null` a 0 ni
       descartar la lectura completa por eso: el resto de los sensores sigue
       siendo valido.
    2. **Se guardan `raw` y `pct` juntos.** El `raw` es la verdad del sensor y no
...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /devices/{device_id}/readings:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    post:
      tags: [ingest]
      summary: Enviar un lote de mediciones
      description: |
        Endpoint que llama la Raspberry Pi. Acepta de 1 a 500 lecturas por
        request. La API key del header debe pertenecer al `device_id` de la ruta.

        Idempotencia: si llega una lectura con el mismo `device_id` + `ts` que
        una ya almacenada, se ignora y se cuenta en `duplicates`. Esto permite
        que la Pi reintente sin miedo a duplicar datos si el POST llego pero la
        respuesta se perdio.

        Validacion parcial: si algunas lecturas del lote son invalidas, las
        validas SI se guardan y las rechazadas se listan en `rejected` con su
        indice. Solo se devuelve 400 si el lote entero es ilegible.

        Reglas de `ts` (v1.2.0):
        - Obligatorio, texto ISO 8601 y **con zona horaria** (`Z` o `-06:00`).
          Sin zona se rechaza: es ambiguo.
        - Se normaliza a UTC con milisegundos: `2026-09-26T18:04:22.123+00:00`.
          El mismo instante escrito en dos zonas cuenta como duplicado.
        - Se rechaza si es anterior a `SMARTGREENAI_MIN_READING_TS`
          (default 2026-01-01T00:00:00Z; una Pi sin hora cree que es 1970) o si
          va mas de `SMARTGREENAI_MAX_CLOCK_SKEW_SECONDS` (default 300) adelante
          del reloj del servidor.

        Motivos de `rejected` (texto `campo: motivo`):
        - `ts: sin zona horaria ...`
        - `ts: debe ser texto ISO 8601 ...` (tambien numeros u horas Unix)
        - `ts: anterior a ...; el reloj del gateway no esta sincronizado`
        - `ts: en el futuro, mas de 300 s adelante del servidor`
        - `ts: field required`
        - `<campo>: <error de validacion>`, por ejemplo `suelo_pct: ensure this
          value is less than or equal to 100`
        - `la lectura no es un objeto JSON`

        Los campos que el contrato no define (hoy `bomba_activa` y `vent_activo`
        del firmware) se ignoran sin rechazar la lectura.
      operationId: ingestReadings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingBatch'
      responses:
        '202':
          description: Lote aceptado para almacenamiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: |
            El `device_id` no esta registrado. La Pi debe darse de alta via
            CONF-04 antes de poder enviar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Lote mayor a 500 lecturas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            El sobre del lote es invalido: `readings` falta, no es arreglo, esta
            vacio o trae mas de 500 elementos. Nota: la implementacion actual
            responde 422 (validacion de FastAPI) tambien en el caso que este
            contrato marca como 413. Las lecturas individuales invalidas NO
            generan 422: van en `rejected`.
  /greenhouses/{greenhouse_id}/readings:
    parameters:
      - name: greenhouse_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [readings]
      summary: Consultar mediciones de un vivero
      description: Lo que consume el dashboard web para las graficas historicas.
      operationId: listReadings
      parameters:
        - name: from
          in: query
          description: Inicio del rango, ISO 8601 UTC
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Fin del rango, ISO 8601 UTC
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        '200':
          description: Mediciones ordenadas por `ts` descendente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoredReading'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Sprint 2: sesiones de usuario (AUTH-01 / AUTH-02) ─────────────────
...
  /greenhouses:
    get:
      tags: [readings]
      summary: Viveros que el usuario puede consultar
      description: |
        Lista los viveros con al menos un dispositivo registrado. Un `producer`
        recibe solo su vivero asignado (o una lista vacia si no tiene);
        `admin` y `super_admin` reciben todos. Lo usa el dashboard para que un
        Administrador, que no tiene vivero asignado, elija cual ver.
      operationId: listGreenhouses
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Background'
      responses:
        '200':
          description: Viveros visibles para el usuario, ordenados por id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GreenhouseSummary'
        '401':
          $ref: '#/components/responses/SessionRequired'
  /greenhouses/{greenhouse_id}/readings/latest:
    parameters:
      - name: greenhouse_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/Background'
    get:
      tags: [readings]
      summary: Ultimas lecturas de un vivero
      description: |
        Ultimo valor NO nulo de cada medicion con unidad, area, antiguedad y
        bandera `stale` (mas vieja que `SMARTGREENAI_READING_STALE_MINUTES`,
        15 por defecto). `sensor_down` indica que la lectura mas reciente del
        vivero llego con ese sensor en `null`. Un `producer` solo consulta su
        vivero asignado. Si el vivero no tiene lecturas responde
        `has_data: false`.

        Es lo que consumen los widgets del dashboard, que la piden cada 15 s
        con `X-SG-Background: 1`.
      operationId: getLatestReadings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Estado actual del vivero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestReadings'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Key emitida al registrar el dispositivo. Para el prototipo es
        suficiente; no es apto para produccion real.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token obtenido en `POST /auth/login`. Solo identifica la sesion; el
        rol se valida contra la BD en cada request.
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      description: Identificador unico del dispositivo, ej. `pi-vivero-01`
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{2,49}$'
    Background:
      name: X-SG-Background
      in: header
      required: false
      description: |
        Mandar `1` en los refrescos automaticos del dashboard: la request se
        valida igual, pero no renueva el plazo de inactividad de la sesion.
      schema:
        type: string
        enum: ['1']
...
  schemas:
...
    ReadingBatch:
      type: object
      required: [readings]
      properties:
        readings:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/Reading'
    Reading:
      type: object
      description: |
        Una lectura del Arduino, enriquecida por la Pi con el timestamp.
        Los nombres de campo son exactamente los que imprime el sketch
        `vivero_sensores.ino` por Serial: la Pi no renombra nada, solo agrega
        `ts` y reenvia.
      required: [ts]
      properties:
        ts:
          type: string
          format: date-time
          description: |
            Momento de la lectura en ISO 8601 con zona horaria. Lo asigna la
            Raspberry Pi al recibir la linea por Serial, NO el Arduino, y solo
            cuando su reloj esta sincronizado por NTP. El backend lo guarda en
            UTC con milisegundos. Sin zona horaria, anterior a 2026-01-01 o mas
            de 300 s en el futuro se rechaza (ver la ingesta).
          example: '2026-09-13T21:04:22.123Z'
        ms:
          type: integer
          format: int64
          description: |
            Milisegundos desde el boot del Arduino. Sirve para detectar
            reinicios (si baja respecto a la lectura anterior, hubo reset).
            No usar para ordenar en el tiempo.
          example: 148022

        temp_c:
          type: number
          format: float
          nullable: true
          description: Temperatura del aire en grados Celsius (DHT11)
          example: 24.2
        hum_aire_pct:
          type: number
          format: float
          nullable: true
          description: Humedad relativa del aire en % (DHT11)
          example: 72.0

        suelo_raw:
          type: integer
          minimum: 0
          maximum: 1023
          nullable: true
          description: 'Lectura ADC cruda del sensor capacitivo. Mas seco = mas alto'
          example: 380
        suelo_pct:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: |
            Humedad de tierra en %, derivada de `suelo_raw` con la calibracion
            vigente. Es un valor relativo a esa calibracion, no humedad
            volumetrica del suelo.
          example: 62

        nivel_raw:
          type: integer
          minimum: 0
          maximum: 1023
          nullable: true
          example: 150
        nivel_pct:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: |
            Nivel del tanque en %. Mide solo el tramo que cubre el sensor, no
            el volumen total del tanque. `0` es un tanque vacio real; sensor
            desconectado llega como `null`.
          example: 75

        luz_raw:
          type: integer
          minimum: 0
          maximum: 1023
          nullable: true
          description: Lectura cruda del sensor de luz, sin calibrar a lux
          example: 306
        luz_nivel:
          type: integer
          minimum: 0
          maximum: 1023
          nullable: true
          description: |
            La misma lectura normalizada a "mas alto = mas luz", para que el
            backend no necesite saber la polaridad del sensor.
          example: 717

        es_dia:
          type: boolean
          description: |
            Dia o noche deducido del propio sensor de luz (el Arduino no tiene
            reloj). El dashboard no debe alertar por falta de sol cuando es
            `false`.
        sol_min_hoy:
          type: integer
          minimum: 0
          description: Minutos de sol pleno acumulados en el dia en curso
          example: 42
        sol_min_prev:
          type: integer
          minimum: 0
          nullable: true
          description: |
            Minutos de sol del dia anterior ya cerrado. `null` mientras no se
            haya completado un ciclo dia-noche desde el ultimo reinicio.
          example: 410

        riego_sugerido:
          type: boolean
          description: |
            El Arduino sugiere regar. Es una recomendacion, no una orden:
            el principio del proyecto es decision support, not autonomy.
        bomba_habilitada:
          type: boolean
          description: |
            `false` cuando el nivel esta en critico o el sensor de nivel fallo.
            Bloqueo de seguridad: la bomba es sumergible y se quema en seco.
        vent:
          type: string
          enum: [BASE, ALTA]
          description: |
            Ventilacion recomendada. `BASE` es la ventilacion constante que el
            cultivo requiere siempre; `ALTA` por calor o exceso de humedad.

        estado:
          type: string
          enum: [OK, AVISO, ALERTA]
          description: |
            Resumen de la lectura. `AVISO` = fuera del rango ideal del rabano;
            `ALERTA` = fuera de limite o sensor caido.
        alertas:
          type: array
          description: |
            Codigos activos en esta lectura. Enum cerrado: si el firmware agrega
            uno nuevo, se actualiza este spec primero.
          items:
            type: string
            enum:
              - TEMP_ALTA
              - TEMP_BAJA
              - TEMP_ALTA_LEVE
              - TEMP_BAJA_LEVE
              - HUM_AIRE_ALTA
              - HUM_AIRE_BAJA
              - HUM_AIRE_ALTA_LEVE
              - HUM_AIRE_BAJA_LEVE
              - SUELO_ENCHARCADO
              - SUELO_SECO
              - SUELO_HUMEDO_LEVE
              - SUELO_SECO_LEVE
              - TANQUE_MITAD
              - TANQUE_UN_CUARTO
              - SOMBRA_PROLONGADA
              - SOL_INSUFICIENTE
              - FALLA_DHT
              - FALLA_SUELO
              - FALLA_NIVEL
          example: [SUELO_SECO, TANQUE_MITAD]
    StoredReading:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
            device_id:
              type: string
            greenhouse_id:
              type: string
            received_at:
              type: string
              format: date-time
              description: |
                Cuando el backend recibio la lectura. Distinto de `ts` cuando
                la Pi estuvo sin internet y sincronizo despues.
        - $ref: '#/components/schemas/Reading'
    IngestResult:
      type: object
      properties:
        accepted:
          type: integer
          description: Lecturas almacenadas
          example: 118
        duplicates:
          type: integer
          description: Ignoradas por `device_id` + `ts` repetido
          example: 2
        rejected:
          type: array
          description: Lecturas invalidas del lote, con su indice en el array
          items:
            type: object
            properties:
              index:
                type: integer
              reason:
                type: string
          example:
            - index: 44
              reason: 'ts: sin zona horaria: agrega Z (UTC) o un desfase como -06:00'
            - index: 45
              reason: 'suelo_pct: ensure this value is less than or equal to 100'
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Codigo estable, legible por maquina
          example: device_not_registered
        message:
          type: string
          description: Explicacion para humanos
          example: 'El dispositivo pi-vivero-01 no esta registrado'

    # ── Sprint 2: usuarios ────────────────────────────────────────────
...
    LatestMetric:
      type: object
      required: [key, label, unit, stale, sensor_down]
      properties:
        key:
          type: string
          enum: [temp_c, hum_aire_pct, suelo_pct, nivel_pct, luz_nivel]
        label:
          type: string
          example: Temperatura del aire
        unit:
          type: string
          example: °C
        value:
          type: number
          nullable: true
          description: Ultimo valor NO nulo. `null` si nunca ha llegado uno.
          example: 24.2
        raw:
          type: integer
          nullable: true
          description: Lectura ADC cruda del mismo momento (suelo, nivel y luz)
        ts:
          type: string
          format: date-time
          nullable: true
        age_seconds:
          type: integer
          nullable: true
        stale:
          type: boolean
        sensor_down:
          type: boolean
        device_id:
          type: string
          nullable: true
        device_location:
          type: string
          nullable: true
        area:
          type: string
          description: |
            Area del vivero a la que pertenece la medicion. El prototipo es un
            solo ambiente, asi que hoy siempre vale
            `SMARTGREENAI_DEFAULT_AREA` (`general`, igual que los actuadores).
            Cuando exista el registro de sensores por area (CONF-04) traera
            el area real.
          example: general
    GreenhouseSummary:
      type: object
      required: [greenhouse_id, devices]
      properties:
        greenhouse_id:
          type: string
          example: vivero-rabano-01
        devices:
          type: integer
          description: Dispositivos registrados en el vivero
          example: 1
        last_seen_at:
          type: string
          format: date-time
          nullable: true
          description: Ultima vez que uno de sus dispositivos envio lecturas
    LatestReadingSummary:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
        age_seconds:
          type: integer
        stale:
          type: boolean
        device_id:
          type: string
        estado:
          type: string
          enum: [OK, AVISO, ALERTA]
          nullable: true
        alertas:
          type: array
          items:
            type: string
        riego_sugerido:
          type: boolean
          nullable: true
        bomba_habilitada:
          type: boolean
          nullable: true
        vent:
          type: string
          enum: [BASE, ALTA]
          nullable: true
        es_dia:
          type: boolean
          nullable: true
    LatestReadings:
      type: object
      required: [greenhouse_id, generated_at, stale_after_minutes, has_data, metrics]
      properties:
        greenhouse_id:
          type: string
        generated_at:
          type: string
          format: date-time
        stale_after_minutes:
          type: integer
          example: 15
        has_data:
          type: boolean
        last_reading:
          allOf:
            - $ref: '#/components/schemas/LatestReadingSummary'
          nullable: true
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/LatestMetric'
    ErrorDetail:
      type: object
      description: |
        Formato real de los errores que lanza el backend (FastAPI envuelve el
        error en `detail`). Ver la nota en la descripcion del API.
      properties:
        detail:
          $ref: '#/components/schemas/Error'
  responses:
    BadRequest:
      description: Payload invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: API key ausente, invalida o de otro dispositivo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SessionRequired:
      description: |
        Sin token (`missing_token`), token invalido o vencido (`invalid_token`)
        o sesion terminada / inactiva mas de 30 min (`session_expired`).
        Incluye el header `WWW-Authenticate: Bearer`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
    Forbidden:
      description: El rol del usuario no tiene permiso (`forbidden`). La accion no se ejecuta.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
...